requirements.txt
plaid-python==18.3.0
